Genetic Algorithm Toolbox for MATLAB, v1.2
Soure: http://uos-codem.github.io/GA-Toolbox/
The Genetic Algorithm Toolbox for MATLAB was developed at the Department of Automatic Control and Systems Engineering of The University of Sheffield, UK, in order to make GA's accessible to the control engineer within the framework of an existing computer-aided control system design package. The toolbox was written with the support of a UK SERC grant, and the final version (v1.2) was completed in 1994.